+++ src/shared/formaDate.js
@@ -1,0 +1,3 @@
+export default function formatDate(date, format = "YYYY/MM/DD") {
+  return date.format(format);
+}